Add policy to limit actions + filter on api data

+++ app/Http/Controllers/CampaignController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\InviteCampaign;
 use App\Models\Campaign;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class CampaignController extends Controller
 {
@@ -37,7 +39,7 @@ class CampaignController extends Controller
 
 	public function getEdit(Campaign $campaign)
 	{
-		// TODO only dm
+		$this->authorize('update', $campaign);
 
 		return view(
 			'campaign.edit',
@@ -50,7 +52,7 @@ class CampaignController extends Controller
 
 	public function postEdit(Campaign $campaign, CreateCampaign $request)
 	{
-		// TODO only dm
+		$this->authorize('update', $campaign);
 
 		$campaign->name = $request->get('name');
 
@@ -59,7 +61,7 @@ class CampaignController extends Controller
 
 	public function resetLink(Campaign $campaign)
 	{
-		// TODO only dm
+		$this->authorize('update', $campaign);
 
 		$campaign->generateInviteId();
 		$campaign->save();
@@ -71,7 +73,7 @@ class CampaignController extends Controller
 
 	public function removePlayer(Campaign $campaign, User $user)
 	{
-		// TODO only dm
+		$this->authorize('update', $campaign);
 
 		$campaign->users()->detach($user);
 
@@ -80,7 +82,7 @@ class CampaignController extends Controller
 
 	public function setDm(Campaign $campaign, User $user)
 	{
-		// TODO only dm
+		$this->authorize('update', $campaign);
 
 		$campaign->users()->syncWithoutDetaching([
 			$user->id => ['is_dm' => true],
@@ -91,7 +93,7 @@ class CampaignController extends Controller
 
 	public function unsetDm(Campaign $campaign, User $user)
 	{
-		// TODO only dm
+		$this->authorize('update', $campaign);
 
 		$campaign->users()->syncWithoutDetaching([
 			$user->id => ['is_dm' => false],
@@ -102,6 +104,8 @@ class CampaignController extends Controller
 
 	public function play(Campaign $campaign)
 	{
+		$this->authorize('view', $campaign);
+
 		return view(
 			'campaign.play',
 			[
@@ -113,12 +117,17 @@ class CampaignController extends Controller
 
 	public function get(Campaign $campaign)
 	{
+		$this->authorize('view', $campaign);
+
 		/** @var User $user */
 		$user = $campaign->users()->find(\Auth::id());
+		$isDm = $user->pivot->is_dm;
 
 		$resourcesQuery = $campaign->resources();
 
-		if( !$user->pivot->is_dm )
+		$relations = ['quests', 'quests.steps', 'quests.comments', 'quests.comments.user', 'quests.comments.resource'];
+
+		if(!$isDm)
 		{
 			$resourcesQuery
 				->whereHas(
@@ -128,13 +137,41 @@ class CampaignController extends Controller
 						$query->whereKey(\Auth::id());
 					}
 				);
+
+			$relations = [
+				'quests' => function(HasMany $query){
+					$query
+						->where('is_visible', true)
+						->select(['id', 'campaign_id', 'name'])
+					;
+				},
+				'quests.steps' => function(HasMany $query){
+					$query
+						->where('is_visible', true)
+						->select(['id', 'quest_id', 'name', 'player_content', 'state'])
+					;
+				},
+				'quests.comments' => function(HasMany $query){
+					$query
+						->where('is_visible', true)
+						->select(['id', 'quest_id', 'step_id', 'user_id', 'resource_id', 'player_text', 'type', 'created_at'])
+					;
+				},
+				'quests.comments.user' => function(BelongsTo $query){
+					$query->select(['id', 'name']);
+				},
+				'quests.comments.resource' => function(BelongsTo $query){
+					$query->select(['id', 'name']);
+				},
+			];
 		}
 
 		return [
-			'campaign' => $campaign->load('quests', 'quests.steps', 'quests.comments', 'quests.comments.user', 'quests.comments.resource'),
+			'campaign' => $campaign->load($relations)->only(['id', 'name', 'quests']),
 			'user' => [
 				'id' => $user->id,
-				'isDM' => $user->pivot->is_dm,
+				'name' => $user->name,
+				'isDM' => $isDm,
 			],
 			'resources' => $resourcesQuery->get(),
 		];
